Make Facebook ID a BigInt

+++ db/migrate/20170218214507_create_users.rb
@@ -1,7 +1,7 @@
 class CreateUsers < ActiveRecord::Migration
   def change
     create_table :users do |t|
-      t.integer :id_fb, null: false
+      t.bigint :id_fb, null: false
       t.string :name_first
       t.string :name_last
       t.string :email
